Add files via upload
If one is not well-versed in writing multi-threading code then this becomes a really tricky question for him. This Java question can be tricky even for experienced and senior programmers, who are not really exposed to deadlock and race conditions. Key point here is order, that is if one acquires resources in a particular order and release resources in reverse order one can easily prevent deadlock.